Controllers and Repositories of Entry and User are done

+++ src/main/java/hu/elte/softech/repository/RankingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RankingRepository extends JpaRepository<Ranking,Integer> {
+public interface RankingRepository extends JpaRepository<Ranking,Long> {
 }
